Abstract a function to a new file
